Change dish category form from products to administration app
